pyproject.toml
[project]
name = "cinemind"
version = "0.1.0"
description = "LLM-powered movie recommendation system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "fastapi>=0.135.3",
    "kaggle>=2.0.0",
    "pandas>=3.0.2",
    "psycopg2>=2.9.11",
    "psycopg[binary]>=3.3.3",
    "pydantic>=2.12.5",
    "pydantic-ai>=1.77.0",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
]

[tool.uv]
package = true

[tool.uv.sources]
cinemind = { path = "src/cinemind", editable = true }

[dependency-groups]
dev = ["mypy>=1.20.0", "pytest>=9.0.2", "ruff>=0.15.9"]

[tool.pytest.ini_options]
markers = ["llm: tests that require a live LLM call"]
testpaths = ["tests"]

[build-system]
requires = ["uv_build>=0.11.2,<0.12.0"]
build-backend = "uv_build"
